+++ src/main/java/data/dto/OwnerDTO.java
@@ -6,28 +6,28 @@ import entity.Car;
 import java.util.List;
 
 public class OwnerDTO {
-    private int id;
+    private long id;
     private String name;
     private List<Car> Cars;
 
     public OwnerDTO() {}
 
-    public OwnerDTO(int id, String name) {
+    public OwnerDTO(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public OwnerDTO(int id, String name, List<Car> cars) {
+    public OwnerDTO(long id, String name, List<Car> cars) {
         this.id = id;
         this.name = name;
         Cars = cars;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
